+++ lib/src/commands/is_in_registry.dart
@@ -28,6 +28,7 @@ class IsInRegistry extends DirCommand<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     await check(directory: directory);
     final messages = <String>[];