+++ lib/src/commands/is_on_pub_dev.dart
@@ -27,6 +27,7 @@ class IsOnPubDev extends DirCommand<bool> {
   Future<bool> exec({
     required Directory directory,
     required GgLog ggLog,
+    Map<String, dynamic> options = const {},
   }) async {
     final messages = <String>[];